add Fenwicktree (cpp)

+++ workbench_cpp/original.cpp
@@ -7,7 +7,7 @@ using matll = vector<vector<ll>>;
 #define min_val(v) *min_element(v.begin(), v.end())
 #define max_val(v) *max_element(v.begin(), v.end())
 #define max_idx(v) distance(v.begin(), max_element(v.begin(), v.end()))
-#define sum(v) accumulate(v.begin(), v.end(), ll(0))
+#define sum(v) accumulate(v.begin(), v.end(), 0LL)
 #define sort_all(v) sort(v.begin(), v.end())
 #define rep(i, n) for (ll i = 0; i < ll(n); i++)
 #define for_itr(id, itr) for (auto& id : itr)
